fix chunk generator reset on forge
